requirements.txt
django
channels-redis
djangorestframework
django-cors-headers
djangorestframework-simplejwt
whitenoise